+++ compute/clusters.go
@@ -6,6 +6,7 @@ import (
 	"log"
 	"sort"
 	"strings"
+	"sync"
 	"time"
 
 	"github.com/databrickslabs/databricks-terraform/common"
@@ -293,8 +294,16 @@ func (a ClustersAPI) ListNodeTypes() (l NodeTypeList, err error) {
 	return
 }
 
+// getOrCreateClusterMutex guards "mounting" cluster creation to prevent multiple
+// redundant instances created at the same name. Compute package private property.
+// https://github.com/databrickslabs/terraform-provider-databricks/issues/445
+var getOrCreateClusterMutex sync.Mutex
+
 // GetOrCreateRunningCluster creates an autoterminating cluster if it doesn't exist
 func (a ClustersAPI) GetOrCreateRunningCluster(name string, custom ...Cluster) (c ClusterInfo, err error) {
+	getOrCreateClusterMutex.Lock()
+	defer getOrCreateClusterMutex.Unlock()
+
 	if len(custom) > 1 {
 		err = fmt.Errorf("You can only specify 1 custom cluster conf, not %d", len(custom))
 		return